add one commen
Now define a release process type in the proc file, as well as a command to run. 
The release command is run when a new release is created. Add the following line at the end of the file: release: echo "You created a new release on Heroku. Here's a cookie! 🍪"